+++ src/hybridmap/NeutralAtomArchitecture.cpp
@@ -11,14 +11,14 @@
 #include "hybridmap/NeutralAtomArchitecture.hpp"
 
 #include "datastructures/SymmetricMatrix.hpp"
-#include "hybridmap/AodOperation.hpp"
 #include "hybridmap/NeutralAtomDefinitions.hpp"
-#include "hybridmap/NeutralAtomOperation.hpp"
 #include "hybridmap/NeutralAtomUtils.hpp"
 #include "ir/Definitions.hpp"
 #include "ir/operations/OpType.hpp"
 #include "ir/operations/Operation.hpp"
-#include "na/computation/entities/Location.hpp"
+#include "na/ir/entities/Location.hpp"
+#include "na/ir/operations/AodOperation.hpp"
+#include "na/ir/operations/NAOpType.hpp"
 
 #include <algorithm>
 #include <cmath>
@@ -98,11 +98,11 @@ void NeutralAtomArchitecture::loadJson(const std::string& filename) {
     ensureGateWithFallback(gateAverageFidelities, "cz", "none");
     ensureGateWithFallback(gateAverageFidelities, "h", "none");
     parameters.gateAverageFidelities = gateAverageFidelities;
-    std::map<NeutralAtomOperationKind, qc::fp> shuttlingTimes;
+    std::map<NAOpType, qc::fp> shuttlingTimes;
 
     for (const auto& [key, value] :
          jsonDataParameters["shuttlingTimes"].items()) {
-      shuttlingTimes.emplace(neutralAtomOperationKindFromString(key), value);
+      shuttlingTimes.emplace(naOpTypeFromString(key), value);
     }
     // compute values for SWAP gate
     qc::fp const swapGateTime =
@@ -133,11 +133,10 @@ void NeutralAtomArchitecture::loadJson(const std::string& filename) {
     }
 
     parameters.shuttlingTimes = shuttlingTimes;
-    std::map<NeutralAtomOperationKind, qc::fp> shuttlingAverageFidelities;
+    std::map<NAOpType, qc::fp> shuttlingAverageFidelities;
     for (const auto& [key, value] :
          jsonDataParameters["shuttlingAverageFidelities"].items()) {
-      shuttlingAverageFidelities.emplace(
-          neutralAtomOperationKindFromString(key), value);
+      shuttlingAverageFidelities.emplace(naOpTypeFromString(key), value);
     }
     parameters.shuttlingAverageFidelities = shuttlingAverageFidelities;
 
@@ -273,22 +272,21 @@ std::string NeutralAtomArchitecture::getAnimationMachine(
   }
   std::string animationMachine = "name: \"Hybrid_" + name + "\"\n";
 
-  animationMachine +=
-      "movement {\n\tmax_speed: " +
-      std::to_string(getShuttlingTime(NeutralAtomOperationKind::AodMove) *
-                     shuttlingSpeedFactor) +
-      "\n}\n";
+  animationMachine += "movement {\n\tmax_speed: " +
+                      std::to_string(getShuttlingTime(NAOpType::AodMove) *
+                                     shuttlingSpeedFactor) +
+                      "\n}\n";
 
-  animationMachine +=
-      "time {\n\tload: " +
-      std::to_string(getShuttlingTime(NeutralAtomOperationKind::AodActivate) /
-                     shuttlingSpeedFactor) +
-      "\n\tstore: " +
-      std::to_string(getShuttlingTime(NeutralAtomOperationKind::AodDeactivate) /
-                     shuttlingSpeedFactor) +
-      "\n\trz: " + std::to_string(getGateTime("x")) +
-      "\n\try: " + std::to_string(getGateTime("x")) +
-      "\n\tcz: " + std::to_string(getGateTime("cz")) + "\n\tunit: \"us\"\n}\n";
+  animationMachine += "time {\n\tload: " +
+                      std::to_string(getShuttlingTime(NAOpType::AodActivate) /
+                                     shuttlingSpeedFactor) +
+                      "\n\tstore: " +
+                      std::to_string(getShuttlingTime(NAOpType::AodDeactivate) /
+                                     shuttlingSpeedFactor) +
+                      "\n\trz: " + std::to_string(getGateTime("x")) +
+                      "\n\try: " + std::to_string(getGateTime("x")) +
+                      "\n\tcz: " + std::to_string(getGateTime("cz")) +
+                      "\n\tunit: \"us\"\n}\n";
 
   animationMachine +=
       "distance {\n\tinteraction: " +
@@ -325,25 +323,21 @@ qc::fp NeutralAtomArchitecture::getOpTime(const qc::Operation* op) const {
   if (op == nullptr) {
     throw std::invalid_argument("Operation must not be null.");
   }
-  const auto* neutralAtomOperation =
-      dynamic_cast<const NeutralAtomOperation*>(op);
-  if (neutralAtomOperation != nullptr &&
-      (neutralAtomOperation->getKind() ==
-           NeutralAtomOperationKind::AodActivate ||
-       neutralAtomOperation->getKind() ==
-           NeutralAtomOperationKind::AodDeactivate)) {
-    return getShuttlingTime(neutralAtomOperation->getKind());
+  const auto naOpType = getNAOpType(*op);
+  if (naOpType == NAOpType::AodActivate ||
+      naOpType == NAOpType::AodDeactivate) {
+    return getShuttlingTime(*naOpType);
   }
-  if (neutralAtomOperation != nullptr &&
-      neutralAtomOperation->getKind() == NeutralAtomOperationKind::AodMove) {
-    const auto v =
-        parameters.shuttlingTimes.at(NeutralAtomOperationKind::AodMove);
+  if (naOpType == NAOpType::AodMove) {
+    const auto v = parameters.shuttlingTimes.at(NAOpType::AodMove);
     const auto* const opAodMove = dynamic_cast<const AodOperation*>(op);
     if (opAodMove == nullptr) {
       throw std::logic_error("An AOD move must be backed by AodOperation.");
     }
-    const auto distanceX = opAodMove->getMaxDistance(Dimension::X);
-    const auto distanceY = opAodMove->getMaxDistance(Dimension::Y);
+    const auto distanceX =
+        opAodMove->getMaxDistance(AodOperation::Dimension::X);
+    const auto distanceY =
+        opAodMove->getMaxDistance(AodOperation::Dimension::Y);
     return (distanceX + distanceY) / v;
   }
   std::string opName;
@@ -370,15 +364,10 @@ qc::fp NeutralAtomArchitecture::getOpFidelity(const qc::Operation* op) const {
   if (op == nullptr) {
     throw std::invalid_argument("Operation must not be null.");
   }
-  const auto* neutralAtomOperation =
-      dynamic_cast<const NeutralAtomOperation*>(op);
-  if (neutralAtomOperation != nullptr &&
-      (neutralAtomOperation->getKind() ==
-           NeutralAtomOperationKind::AodActivate ||
-       neutralAtomOperation->getKind() ==
-           NeutralAtomOperationKind::AodDeactivate ||
-       neutralAtomOperation->getKind() == NeutralAtomOperationKind::AodMove)) {
-    return getShuttlingAverageFidelity(neutralAtomOperation->getKind());
+  const auto naOpType = getNAOpType(*op);
+  if (naOpType == NAOpType::AodActivate ||
+      naOpType == NAOpType::AodDeactivate || naOpType == NAOpType::AodMove) {
+    return getShuttlingAverageFidelity(*naOpType);
   }
   std::string opName;
   const auto nQubits = op->getNqubits();
@@ -391,15 +380,9 @@ qc::fp NeutralAtomArchitecture::getOpFidelity(const qc::Operation* op) const {
 
 std::set<CoordIndex>
 NeutralAtomArchitecture::getBlockedCoordIndices(const qc::Operation* op) const {
-  const auto* neutralAtomOperation =
-      dynamic_cast<const NeutralAtomOperation*>(op);
-  if (op->getNqubits() == 1 || (neutralAtomOperation != nullptr &&
-                                (neutralAtomOperation->getKind() ==
-                                     NeutralAtomOperationKind::AodActivate ||
-                                 neutralAtomOperation->getKind() ==
-                                     NeutralAtomOperationKind::AodDeactivate ||
-                                 neutralAtomOperation->getKind() ==
-                                     NeutralAtomOperationKind::AodMove))) {
+  const auto naOpType = getNAOpType(*op);
+  if (op->getNqubits() == 1 || naOpType == NAOpType::AodActivate ||
+      naOpType == NAOpType::AodDeactivate || naOpType == NAOpType::AodMove) {
     return op->getUsedQubits();
   }
   std::set<CoordIndex> blockedCoordIndices;