+++ src/na/nasp/CodeGenerator.cpp
@@ -14,14 +14,14 @@
 #include "ir/Definitions.hpp"
 #include "ir/QuantumComputation.hpp"
 #include "ir/operations/OpType.hpp"
-#include "na/computation/NAComputation.hpp"
-#include "na/computation/entities/Atom.hpp"
-#include "na/computation/operations/GlobalCZOp.hpp"
-#include "na/computation/operations/GlobalRYOp.hpp"
-#include "na/computation/operations/LoadOp.hpp"
-#include "na/computation/operations/LocalRZOp.hpp"
-#include "na/computation/operations/MoveOp.hpp"
-#include "na/computation/operations/StoreOp.hpp"
+#include "na/ir/NAComputation.hpp"
+#include "na/ir/entities/Atom.hpp"
+#include "na/ir/operations/NAComputationGlobalCZOperation.hpp"
+#include "na/ir/operations/NAComputationGlobalRYOperation.hpp"
+#include "na/ir/operations/NAComputationLoadOperation.hpp"
+#include "na/ir/operations/NAComputationLocalRZOperation.hpp"
+#include "na/ir/operations/NAComputationMoveOperation.hpp"
+#include "na/ir/operations/NAComputationStoreOperation.hpp"
 #include "na/nasp/Solver.hpp"
 
 #include <algorithm>
@@ -117,42 +117,38 @@ auto CodeGenerator::generate(const QuantumComputation& input,
       code.emplaceInitialLocation(atom, pos);
     }
     if (!loadAtoms.empty()) {
-      code.emplaceBack<LoadOp>(loadAtoms);
+      code.emplaceBack<NAComputationLoadOperation>(loadAtoms);
     }
   }
   const auto ops = layer.getExecutablesOfType(H, 0);
   std::unordered_set<Qubit> affectedQubits;
-  std::transform(ops.cbegin(), ops.cend(),
-                 std::inserter(affectedQubits, affectedQubits.end()),
-                 [](const auto& v) -> auto {
-                   return v->getOperation()->getTargets().front();
-                 });
+  std::transform(
+      ops.cbegin(), ops.cend(),
+      std::inserter(affectedQubits, affectedQubits.end()),
+      [](const auto& v) { return v->getOperation()->getTargets().front(); });
   if (affectedQubits.size() != flattened.getNqubits() ||
       ops.size() != affectedQubits.size()) {
     throw std::invalid_argument("Not all atoms are initialized to plus state.");
   }
   // initialize atoms in to |+> state starting in |0> state
-  code.emplaceBack<GlobalRYOp>(globalZone, PI_2);
-  std::for_each(ops.cbegin(), ops.cend(),
-                [](const auto& v) -> void { v->execute(); });
+  code.emplaceBack<NAComputationGlobalRYOperation>(globalZone, PI_2);
+  std::for_each(ops.cbegin(), ops.cend(), [](const auto& v) { v->execute(); });
   // Reference to the executable set of the input circuit
   const auto& executableSet = layer.getExecutableSet();
   if (result.stages.front().rydberg) {
-    code.emplaceBack<GlobalCZOp>(interactionZone);
+    code.emplaceBack<NAComputationGlobalCZOperation>(interactionZone);
     // find and execute corresponding gates in input circuit
     for (const auto& g : result.stages.front().gates) {
-      const auto& it =
-          std::find_if(executableSet.begin(), executableSet.end(),
-                       [g](const auto& v) -> bool {
-                         if (v->getOperation()->getType() == Z &&
-                             v->getOperation()->getNcontrols() == 1) {
-                           const auto& usedQubits =
-                               v->getOperation()->getUsedQubits();
-                           const auto [first, second] = g.qubits;
-                           return std::set{first, second} == usedQubits;
-                         }
-                         return false;
-                       });
+      const auto& it = std::find_if(
+          executableSet.begin(), executableSet.end(), [g](const auto& v) {
+            if (v->getOperation()->getType() == Z &&
+                v->getOperation()->getNcontrols() == 1) {
+              const auto& usedQubits = v->getOperation()->getUsedQubits();
+              const auto [first, second] = g.qubits;
+              return std::set{first, second} == usedQubits;
+            }
+            return false;
+          });
       if (it == executableSet.end()) {
         throw std::invalid_argument(
             "Gate in input circuit has no correspondence in solution.");
@@ -184,31 +180,29 @@ auto CodeGenerator::generate(const QuantumComputation& input,
       wasAOD[i] = q.a;
     }
     if (!storeAtoms.empty()) {
-      code.emplaceBack<StoreOp>(storeAtoms);
+      code.emplaceBack<NAComputationStoreOperation>(storeAtoms);
     }
     if (!loadAtoms.empty()) {
-      code.emplaceBack<LoadOp>(loadAtoms);
+      code.emplaceBack<NAComputationLoadOperation>(loadAtoms);
     }
     if (!moveAtoms.empty()) {
-      code.emplaceBack<MoveOp>(moveAtoms, targetLocations);
+      code.emplaceBack<NAComputationMoveOperation>(moveAtoms, targetLocations);
     }
     if (result.stages.at(t).rydberg) {
-      code.emplaceBack<GlobalCZOp>(interactionZone);
+      code.emplaceBack<NAComputationGlobalCZOperation>(interactionZone);
     }
     // find and execute corresponding gates in input circuit
     for (const auto& g : result.stages.at(t).gates) {
-      const auto& it =
-          std::find_if(executableSet.begin(), executableSet.end(),
-                       [g](const auto& v) -> bool {
-                         if (v->getOperation()->getType() == Z &&
-                             v->getOperation()->getNcontrols() == 1) {
-                           const auto& usedQubits =
-                               v->getOperation()->getUsedQubits();
-                           const auto [first, second] = g.qubits;
-                           return std::set{first, second} == usedQubits;
-                         }
-                         return false;
-                       });
+      const auto& it = std::find_if(
+          executableSet.begin(), executableSet.end(), [g](const auto& v) {
+            if (v->getOperation()->getType() == Z &&
+                v->getOperation()->getNcontrols() == 1) {
+              const auto& usedQubits = v->getOperation()->getUsedQubits();
+              const auto [first, second] = g.qubits;
+              return std::set{first, second} == usedQubits;
+            }
+            return false;
+          });
       if (it == executableSet.end()) {
         throw std::invalid_argument(
             "Gate in input circuit has no correspondence in solution.");
@@ -217,7 +211,7 @@ auto CodeGenerator::generate(const QuantumComputation& input,
     }
   }
   if (!executableSet.empty()) {
-    code.emplaceBack<GlobalRYOp>(globalZone, -PI_4);
+    code.emplaceBack<NAComputationGlobalRYOperation>(globalZone, -PI_4);
     while (!executableSet.empty()) {
       const auto& v = (*executableSet.cbegin());
       if (v->getOperation()->getType() != H) {
@@ -225,10 +219,10 @@ auto CodeGenerator::generate(const QuantumComputation& input,
             "Not all non CZ-gates in input circuit are executed.");
       }
       const auto q = v->getOperation()->getTargets().front();
-      code.emplaceBack<LocalRZOp>(*atoms[q], PI);
+      code.emplaceBack<NAComputationLocalRZOperation>(*atoms[q], PI);
       v->execute();
     }
-    code.emplaceBack<GlobalRYOp>(globalZone, PI_4);
+    code.emplaceBack<NAComputationGlobalRYOperation>(globalZone, PI_4);
   }
   return code;
 }